requirements.txt
torch>=2.1
torchaudio>=2.1
PyYAML>=6.0
numpy>=1.26
pytest>=8.0
